+++ tests/EnumTypeHintTest.php
@@ -9,8 +9,8 @@ class EnumTypeHintTest extends TestCase
 {
     public function test_can_pass_an_enum_instance_to_a_type_hinted_method()
     {
-        $userType1 = UserType::getInstance(UserType::SuperAdministrator);
-        $userType2 = UserType::getInstance(UserType::Moderator);
+        $userType1 = UserType::fromValue(UserType::SuperAdministrator);
+        $userType2 = UserType::fromValue(UserType::Moderator);
 
         $this->assertTrue($this->typeHintedMethod($userType1));
         $this->assertFalse($this->typeHintedMethod($userType2));